requirements.txt
about-time==4.2.1
alive-progress==3.1.5
altgraph==0.17.4
anyio==4.2.0
art==6.1
asyncio==3.4.3
auto-py-to-exe==2.42.0
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
clr-loader==0.2.6
cocos2d==0.6.10
colorama==0.4.6
comtypes==1.3.0
customtkinter==5.2.2
Cython==0.29.37
darkdetect==0.8.0
DateTime==5.4
dearpygui==1.10.1
decorator==4.4.2
Eel==0.16.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
glfw==2.6.5
grapheme==0.6.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
imgui==2.0.0
Jinja2==3.0.1
keyboard==0.13.5
MarkupSafe==2.0.1
MouseInfo==0.1.3
moviepy==1.0.3
Nuitka==2.0.2
numpy==1.26.4
ordered-set==4.1.0
packaging==23.2
Panda3D==1.10.14
panda3d-gltf==1.1.0
panda3d-simplepbr==0.11.2
pefile==2023.2.7
pep517==0.13.1
PIL-Tools==1.1.0
pillow==10.2.0
prettytable==3.9.0
proglog==0.1.10
proxy_tools==0.1.0
psutil==5.9.8
PyAutoGUI==0.9.54
pybase64==1.3.2
pycparser==2.21
pycryptodome==3.20.0
pygame==2.5.2
PyGetWindow==0.0.9
pyglet==1.5.28
pyinstaller @ file:///C:/pytools/pyinstaller-6.4.0
pyinstaller-hooks-contrib==2024.1
PyMsgBox==1.0.9
PyOpenGL==3.1.7
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.30
PySDL2==0.9.16
pythonnet==3.0.3
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.4
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinstyles==1.6
requests==2.25.1
screeninfo==0.8.1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sv-ttk==2.6.0
threaded==4.2.0
tkhtmlview==0.2.0
tkinterweb==3.23.5
tqdm==4.66.2
ttkthemes==3.2.2
typing_extensions==4.9.0
urllib3==1.26.6
ursina==6.1.2
wcwidth==0.2.13
wheel==0.42.0
whichcraft==0.6.1
winregistry==1.1.1
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
